Pulling out TestInstrumentation into a reusable instrumenter
